requirements.txt
# Streamlit 웹앱 프레임워크
streamlit==1.31.0

# 데이터 처리
pandas==2.2.0
numpy==1.26.0

# 시각화
plotly==5.18.0
matplotlib==3.8.0

# OpenAI API (2단계에서 사용)
openai==1.10.0

# 환경 변수 관리
python-dotenv==1.0.0

# 날짜 처리
python-dateutil==2.8.2

# Google OAuth (3단계에서 사용)
google-auth-oauthlib==1.2.0
google-auth==2.27.0
